backend/requirements.txt
# FastAPI and server
fastapi==0.115.0
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.9.0
pydantic-settings==2.5.2

# Database
sqlmodel==0.0.14
alembic==1.12.1

# Security
python-jose[cryptography]==3.3.0
passlib[argon2]==1.7.4
argon2-cffi==23.1.0

# LDAP (for Active Directory integration)
ldap3==2.9.1

# Email
python-dotenv==1.0.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2

# Code quality
pylint==3.0.3
black==23.12.1
